[Lightning WIP] Split up different trainers
