+++ src/gnn_tracking/test_training_integration.py
@@ -14,7 +14,7 @@ from gnn_tracking.models.track_condensation_networks import (
     PreTrainedECGraphTCN,
 )
 from gnn_tracking.postprocessing.dbscanscanner import dbscan_scan
-from gnn_tracking.training.tcn_trainer import TCNTrainer
+from gnn_tracking.training.tc import TCNTrainer
 from gnn_tracking.utils.seeds import fix_seeds
